requirements.txt
# Runtime dependencies
pyyaml>=6.0
jsonschema>=4.0

# Test dependencies
pytest>=7.0
pytest-cov>=4.0

# Security scanning (used in CI pipeline)
bandit>=1.7
